Player Sprites (More to come but this is a good start)
We have the children in the game.

When using these sprites, ensure that the FILTER MODE is POINT for each sprite. I've already set it, but it may revert. When the Filter Mode is not point (can be bilinear or trilinear), it makes the sprite look very foggy and blurry because these are very small sprites (theyre like 40x40 pixels)

Here is default, punching, and block sprites. Need to add taking damage sprite, walking sprites, and maybe jumping. Once we have all these sprites, we can use the Unity In-House animator. Gaming. Poggers.