Just General Refactoring
Renamed a lot of scripts and scenes to make them more cohesive; also did this so we can load different "Minigames" when more are made (naming a script just "Minigame" irked me).

Added a small part in  Countdown timer to let us randomize which minigame is selected.
This also needs to be added in the code of the Player death. I think I might make a "minigame and general data manager object" for this that doesnt get destroyed on scene load. With this potential object, we can also keep track of  which player has the buff, and current score.

Players are still unable to take damage, but once they can the main way that a minigame is loaded is the player death. Ill be working on this in my own time I feel.

In general, this is not an important commit.